1.6.0

+++ hexoweb/functions.py
@@ -1,9 +1,9 @@
 import os
-from core.settings import ALL_SETTINGS
+from core.QexoSettings import ALL_SETTINGS
 import requests
 from django.template.defaulttags import register
-from core.settings import QEXO_VERSION
-from .models import Cache, SettingModel, FriendModel, NotificationModel
+from core.QexoSettings import QEXO_VERSION
+from .models import Cache, SettingModel, FriendModel, NotificationModel, CustomModel
 import github
 import json
 import boto3
@@ -16,6 +16,8 @@ from zlib import crc32 as zlib_crc32
 from urllib.parse import quote
 from time import strftime, localtime
 import tarfile
+from ftplib import FTP
+from onepush import notify
 
 disable_warnings()
 
@@ -329,7 +331,8 @@ def update_configs_cache(s=None):
 def delete_all_caches():
     caches = Cache.objects.all()
     for cache in caches:
-        cache.delete()
+        if cache.name != "update":
+            cache.delete()
 
 
 def delete_posts_caches():
@@ -367,13 +370,47 @@ def save_setting(name, content):
     return new_set
 
 
+def save_custom(name, content):
+    obj = CustomModel.objects.filter(name=name)
+    if obj.count() == 1:
+        obj.delete()
+    if obj.count() > 1:
+        for i in obj:
+            i.delete()
+    new_set = CustomModel()
+    new_set.name = str(name)
+    if content is not None:
+        new_set.content = str(content)
+    else:
+        new_set.content = ""
+    new_set.save()
+    return new_set
+
+
+def save_cache(name, content):
+    obj = Cache.objects.filter(name=name)
+    if obj.count() == 1:
+        obj.delete()
+    if obj.count() > 1:
+        for i in obj:
+            i.delete()
+    new_set = Cache()
+    new_set.name = str(name)
+    if content is not None:
+        new_set.content = str(content)
+    else:
+        new_set.content = ""
+    new_set.save()
+    return new_set
+
+
 def upload_to_s3(file, key_id, access_key, endpoint_url, bucket, path, prev_url):
     # 处理 path
     now = date.today()
     photo_stream = file.read()
     path = path.replace("{year}", str(now.year)).replace("{month}", str(now.month)).replace("{day}",
                                                                                             str(now.day)) \
-        .replace("{filename}", file.name).replace("{time}", str(time())) \
+        .replace("{filename}", file.name[0:-len(file.name.split(".")[-1]) - 1]).replace("{time}", str(time())) \
         .replace("{extName}", file.name.split(".")[-1]).replace("{md5}",
                                                                 md5(photo_stream).hexdigest())
 
@@ -388,6 +425,53 @@ def upload_to_s3(file, key_id, access_key, endpoint_url, bucket, path, prev_url)
     bucket.put_object(Key=path, Body=photo_stream, ContentType=file.content_type)
 
     return prev_url + "/" + path
+
+
+def upload_to_custom(file, api, post_params, json_path, custom_body, custom_header, custom_url):
+    if custom_header:
+        if custom_body:
+            response = requests.post(api, data=json.loads(custom_body),
+                                     headers=json.loads(custom_header),
+                                     files={post_params: [file.name, file.read(),
+                                                          file.content_type]})
+        else:
+            response = requests.post(api, data={}, headers=json.loads(custom_header),
+                                     files={post_params: [file.name, file.read(),
+                                                          file.content_type]})
+    else:
+        if custom_body:
+            response = requests.post(api, data=json.loads(custom_body),
+                                     files={post_params: [file.name, file.read(),
+                                                          file.content_type]})
+        else:
+            response = requests.post(api, data={},
+                                     files={post_params: [file.name, file.read(),
+                                                          file.content_type]})
+    if json_path:
+        json_path = json_path.split(".")
+        response.encoding = "utf8"
+        data = response.json()
+        for path in json_path:
+            data = data[path]
+    else:
+        data = response.text
+    return str(custom_url) + data
+
+
+def upload_to_ftp(file, host, port, user, password, path, prev_url):
+    ftp = FTP()
+    ftp.set_debuglevel(0)
+    ftp.encoding = 'UTF8'
+    ftp.connect(host, int(port))
+    ftp.login(user, password)
+    now = date.today()
+    path = path.replace("{year}", str(now.year)).replace("{month}", str(now.month)).replace("{day}",
+                                                                                            str(now.day)) \
+        .replace("{filename}", file.name[0:-len(file.name.split(".")[-1]) - 1]).replace("{time}", str(time())) \
+        .replace("{extName}", file.name.split(".")[-1])
+    bufsize = 1024
+    ftp.storbinary('STOR ' + path, file, bufsize)
+    return prev_url + path
 
 
 def get_latest_version():
@@ -668,7 +752,11 @@ def CreateNotification(label, content, now):
     N.content = content
     N.time = str(float(now))
     N.save()
-    return True
+    try:
+        notify_me(label, content)
+    except:
+        pass
+    return N
 
 
 def GetNotifications():
@@ -687,4 +775,13 @@ def GetNotifications():
 def DelNotification(_time):
     N = NotificationModel.objects.get(time=_time)
     N.delete()
-    return True
+    return N
+
+
+def notify_me(title, content):
+    config = SettingModel.objects.get(name="ONEPUSH").content
+    if config:
+        config = json.loads(config)
+    else:
+        return False
+    return notify(config["notifier"], **config["params"], title="Qexo消息: "+title, content=content).text
